Slightly less hacky way of doing as the previous two changes
Since we need to force-forward the event to the view anyway, we
better do it to the window's own view (which keeps a sensible state)
and forward it to the QWindow from -[QNSView handleMouseEvent:].

This saves us from having to clone the event and touching QNSView
internal mouse button state. It also avoids disturbing the Cocoa
event flow by sending an NSEvent to some random (to Cocoa) window.

Change-Id: Ie81545b6f535ccf381c3871f6463d9f2f2a4f3c1